server/requirements.txt
fastapi==0.104.1
pymongo==4.6.0
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
